LPS-190119 - move Configuration and Util classes to api module and refactor
