+++ modules/dxp/apps/scim/scim-configuration-web/build.gradle
@@ -26,5 +26,5 @@ dependencies {
 	compileOnly project(":core:petra:petra-lang")
 	compileOnly project(":core:petra:petra-reflect")
 	compileOnly project(":core:petra:petra-string")
-	compileOnly project(":dxp:apps:scim:scim-api")
+	compileOnly project(":dxp:apps:scim:scim-rest-api")
 }
