+++ modules/dxp/apps/scim/scim-rest-api/bnd.bnd
@@ -2,5 +2,7 @@ Bundle-Name: Liferay SCIM REST API
 Bundle-SymbolicName: com.liferay.scim.rest.api
 Bundle-Version: 1.0.2
 Export-Package:\
+	com.liferay.scim.rest.configuration,\
 	com.liferay.scim.rest.dto.v1_0,\
-	com.liferay.scim.rest.resource.v1_0
+	com.liferay.scim.rest.resource.v1_0,\
+	com.liferay.scim.rest.util
