+++ modules/dxp/apps/scim/scim-rest-impl/src/main/java/com/liferay/scim/rest/internal/configuration/ScimClientOAuth2ApplicationConfigurationFactory.java
@@ -29,8 +29,9 @@ import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.util.HashMapDictionaryBuilder;
 import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.util.PropsValues;
+import com.liferay.scim.rest.configuration.ScimClientOAuth2ApplicationConfiguration;
 import com.liferay.scim.rest.internal.provider.ScimClientBearerTokenProvider;
-import com.liferay.scim.rest.internal.util.ScimClientUtil;
+import com.liferay.scim.rest.util.ScimClientUtil;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -52,7 +53,7 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
  * @author Olivér Kecskeméty
  */
 @Component(
-	configurationPid = "com.liferay.scim.rest.internal.configuration.ScimClientOAuth2ApplicationConfiguration",
+	configurationPid = "com.liferay.scim.rest.configuration.ScimClientOAuth2ApplicationConfiguration",
 	configurationPolicy = ConfigurationPolicy.REQUIRE, service = {}
 )
 public class ScimClientOAuth2ApplicationConfigurationFactory {
